fix(compilation): Fixed compilation warnings and undefined behaviors detected by ubsan
 - Add virtual destructor to GameInteractionEffectBase to prevent UB when deleting through a base pointer.
 - Use UINT32_MAX instead of -1
 - Many other fixes

+++ soh/soh/Enhancements/debugger/actorViewer.cpp
@@ -920,7 +920,7 @@ void ActorViewerWindow::DrawElement() {
 
         PushStyleCombobox(THEME_COLOR);
         if (ImGui::BeginCombo("Actor Type", acMapping[category])) {
-            for (int i = 0; i < acMapping.size(); i++) {
+            for (size_t i = 0; i < acMapping.size(); i++) {
                 if (ImGui::Selectable(acMapping[i])) {
                     category = i;
                     PopulateActorDropdown(category, list);
@@ -935,7 +935,7 @@ void ActorViewerWindow::DrawElement() {
         }
 
         if (ImGui::BeginCombo("Actor", filler.c_str())) {
-            for (int i = 0; i < list.size(); i++) {
+            for (size_t i = 0; i < list.size(); i++) {
                 std::string label = std::to_string(i) + ": " + ActorDB::Instance->RetrieveEntry(list[i]->id).name;
                 std::string description = GetActorDescription(list[i]->id);
                 if (description != "")
